Extracting node validator functions to hooks
