+++ src/layout/Address/index.tsx
@@ -2,14 +2,14 @@ import React, { forwardRef } from 'react';
 import type { JSX } from 'react';
 
 import { useDisplayData } from 'src/features/displayData/useDisplayData';
-import { FrontendValidationSource, ValidationMask } from 'src/features/validation';
 import { AddressComponent } from 'src/layout/Address/AddressComponent';
 import { AddressSummary } from 'src/layout/Address/AddressSummary/AddressSummary';
 import { AddressDef } from 'src/layout/Address/config.def.generated';
+import { useAddressValidation } from 'src/layout/Address/useAddressValidation';
 import { SummaryItemSimple } from 'src/layout/Summary/SummaryItemSimple';
 import type { LayoutValidationCtx } from 'src/features/devtools/layoutValidation/types';
 import type { DisplayDataProps } from 'src/features/displayData';
-import type { ComponentValidation, ValidationDataSources } from 'src/features/validation';
+import type { ComponentValidation } from 'src/features/validation';
 import type { PropsFromGenericComponent, ValidateComponent } from 'src/layout';
 import type { SummaryRendererProps } from 'src/layout/LayoutComponent';
 import type { Summary2Props } from 'src/layout/Summary2/SummaryComponent2/types';
@@ -39,50 +39,8 @@ export class Address extends AddressDef implements ValidateComponent<'Address'> 
     return false;
   }
 
-  runComponentValidation(
-    node: LayoutNode<'Address'>,
-    { formDataSelector, nodeDataSelector }: ValidationDataSources,
-  ): ComponentValidation[] {
-    const dataModelBindings = nodeDataSelector(
-      (picker) => picker(node.id, 'Address')?.layout.dataModelBindings,
-      [node.id],
-    );
-    if (!dataModelBindings) {
-      return [];
-    }
-    const validations: ComponentValidation[] = [];
-
-    const zipCodeField = dataModelBindings.zipCode;
-    const zipCode = zipCodeField ? formDataSelector(zipCodeField) : undefined;
-    const zipCodeAsString = typeof zipCode === 'string' || typeof zipCode === 'number' ? String(zipCode) : undefined;
-
-    // TODO(Validation): Add better message for the special case of 0000 or add better validation for zipCodes that the API says are invalid
-    if (zipCodeAsString && (!zipCodeAsString.match(/^\d{4}$/) || zipCodeAsString === '0000')) {
-      validations.push({
-        message: { key: 'address_component.validation_error_zipcode' },
-        severity: 'error',
-        bindingKey: 'zipCode',
-        source: FrontendValidationSource.Component,
-        category: ValidationMask.Component,
-      });
-    }
-
-    const houseNumberField = dataModelBindings.houseNumber;
-    const houseNumber = houseNumberField ? formDataSelector(houseNumberField) : undefined;
-    const houseNumberAsString =
-      typeof houseNumber === 'string' || typeof houseNumber === 'number' ? String(houseNumber) : undefined;
-
-    if (houseNumberAsString && !houseNumberAsString.match(/^[a-z,A-Z]\d{4}$/)) {
-      validations.push({
-        message: { key: 'address_component.validation_error_house_number' },
-        severity: 'error',
-        bindingKey: 'houseNumber',
-        source: FrontendValidationSource.Component,
-        category: ValidationMask.Component,
-      });
-    }
-
-    return validations;
+  useComponentValidation(node: LayoutNode<'Address'>): ComponentValidation[] {
+    return useAddressValidation(node);
   }
 
   validateDataModelBindings(ctx: LayoutValidationCtx<'Address'>): string[] {
