+++ src/layout/FileUploadWithTag/index.tsx
@@ -1,16 +1,16 @@
 import React, { forwardRef, type JSX } from 'react';
 
-import { isAttachmentUploaded } from 'src/features/attachments';
-import { FrontendValidationSource, ValidationMask } from 'src/features/validation';
 import { AttachmentSummaryComponent2 } from 'src/layout/FileUpload/AttachmentSummaryComponent2';
 import { FileUploadComponent } from 'src/layout/FileUpload/FileUploadComponent';
 import { FileUploadLayoutValidator } from 'src/layout/FileUpload/FileUploadLayoutValidator';
 import { AttachmentSummaryComponent } from 'src/layout/FileUpload/Summary/AttachmentSummaryComponent';
+import { useValidateMinNumberOfAttachments } from 'src/layout/FileUpload/useValidateMinNumberOfAttachments';
 import { FileUploadWithTagDef } from 'src/layout/FileUploadWithTag/config.def.generated';
+import { useValidateMissingTag } from 'src/layout/FileUploadWithTag/useValidateMissingTag';
 import { LayoutPage } from 'src/utils/layout/LayoutPage';
 import type { LayoutValidationCtx } from 'src/features/devtools/layoutValidation/types';
 import type { DisplayDataProps } from 'src/features/displayData';
-import type { AttachmentValidation, ComponentValidation, ValidationDataSources } from 'src/features/validation';
+import type { ComponentValidation } from 'src/features/validation';
 import type { PropsFromGenericComponent, ValidateComponent } from 'src/layout';
 import type { NodeValidationProps } from 'src/layout/layout';
 import type { SummaryRendererProps } from 'src/layout/LayoutComponent';
@@ -47,72 +47,12 @@ export class FileUploadWithTag extends FileUploadWithTagDef implements ValidateC
   }
 
   // This component does not have empty field validation, so has to override its inherited method
-  runEmptyFieldValidation(): ComponentValidation[] {
+  useEmptyFieldValidation(): ComponentValidation[] {
     return [];
   }
 
-  runComponentValidation(
-    node: LayoutNode<'FileUploadWithTag'>,
-    { attachmentsSelector, nodeDataSelector }: ValidationDataSources,
-  ): ComponentValidation[] {
-    const validations: ComponentValidation[] = [];
-    const minNumberOfAttachments = nodeDataSelector(
-      (picker) => picker(node.id, 'FileUploadWithTag')?.item?.minNumberOfAttachments,
-      [node.id],
-    );
-
-    // Validate minNumberOfAttachments
-    const attachments = attachmentsSelector(node.id);
-    if (
-      minNumberOfAttachments !== undefined &&
-      minNumberOfAttachments > 0 &&
-      attachments.length < minNumberOfAttachments
-    ) {
-      validations.push({
-        message: {
-          key: 'form_filler.file_uploader_validation_error_file_number',
-          params: [minNumberOfAttachments],
-        },
-        severity: 'error',
-        source: FrontendValidationSource.Component,
-        // Treat visibility of minNumberOfAttachments the same as required to prevent showing an error immediately
-        category: ValidationMask.Required,
-      });
-    }
-
-    // Validate missing tags
-    for (const attachment of attachments) {
-      if (
-        isAttachmentUploaded(attachment) &&
-        (attachment.data.tags === undefined || attachment.data.tags.length === 0)
-      ) {
-        const tagKey = nodeDataSelector(
-          (picker) => picker(node.id, 'FileUploadWithTag')?.item?.textResourceBindings?.tagTitle,
-          [node.id],
-        );
-        const tagReference = tagKey
-          ? {
-              key: tagKey,
-              makeLowerCase: true,
-            }
-          : 'tag';
-
-        const validation: AttachmentValidation = {
-          message: {
-            key: 'form_filler.file_uploader_validation_error_no_chosen_tag',
-            params: [tagReference],
-          },
-          severity: 'error',
-          source: FrontendValidationSource.Component,
-          attachmentId: attachment.data.id,
-          // Treat visibility of missing tag the same as required to prevent showing an error immediately
-          category: ValidationMask.Required,
-        };
-        validations.push(validation);
-      }
-    }
-
-    return validations;
+  useComponentValidation(node: LayoutNode<'FileUploadWithTag'>): ComponentValidation[] {
+    return [...useValidateMinNumberOfAttachments(node), ...useValidateMissingTag(node)];
   }
 
   isDataModelBindingsRequired(node: LayoutNode<'FileUploadWithTag'>): boolean {
