+++ src/layout/List/index.tsx
@@ -5,15 +5,14 @@ import type { JSONSchema7Definition } from 'json-schema';
 
 import { useDisplayData } from 'src/features/displayData/useDisplayData';
 import { evalQueryParameters } from 'src/features/options/evalQueryParameters';
-import { FrontendValidationSource, ValidationMask } from 'src/features/validation';
 import { ListDef } from 'src/layout/List/config.def.generated';
 import { ListComponent } from 'src/layout/List/ListComponent';
 import { ListSummary } from 'src/layout/List/ListSummary';
+import { useValidateListIsEmpty } from 'src/layout/List/useValidateListIsEmpty';
 import { SummaryItemSimple } from 'src/layout/Summary/SummaryItemSimple';
-import { getFieldNameKey } from 'src/utils/formComponentUtils';
 import type { LayoutValidationCtx } from 'src/features/devtools/layoutValidation/types';
 import type { DisplayDataProps } from 'src/features/displayData';
-import type { ComponentValidation, ValidationDataSources } from 'src/features/validation';
+import type { ComponentValidation } from 'src/features/validation';
 import type { PropsFromGenericComponent } from 'src/layout';
 import type { ExprResolver, SummaryRendererProps } from 'src/layout/LayoutComponent';
 import type { Summary2Props } from 'src/layout/Summary2/SummaryComponent2/types';
@@ -62,63 +61,8 @@ export class List extends ListDef {
     );
   }
 
-  runEmptyFieldValidation(
-    node: LayoutNode<'List'>,
-    { formDataSelector, invalidDataSelector, nodeDataSelector }: ValidationDataSources,
-  ): ComponentValidation[] {
-    const required = nodeDataSelector(
-      (picker) => {
-        const item = picker(node.id, 'List')?.item;
-        return item && 'required' in item ? item.required : false;
-      },
-      [node.id],
-    );
-    const dataModelBindings = nodeDataSelector(
-      (picker) => picker(node.id, 'List')?.layout.dataModelBindings,
-      [node.id],
-    );
-    if (!required || !dataModelBindings) {
-      return [];
-    }
-
-    const references = Object.values(dataModelBindings);
-    const validations: ComponentValidation[] = [];
-    const textResourceBindings = nodeDataSelector(
-      (picker) => picker(node.id, 'List')?.item?.textResourceBindings,
-      [node.id],
-    );
-
-    let listHasErrors = false;
-    for (const reference of references) {
-      const data = formDataSelector(reference) ?? invalidDataSelector(reference);
-      const dataAsString =
-        typeof data === 'string' || typeof data === 'number' || typeof data === 'boolean' ? String(data) : undefined;
-
-      if (!dataAsString?.length) {
-        listHasErrors = true;
-      }
-    }
-    if (listHasErrors) {
-      const key = textResourceBindings?.requiredValidation
-        ? textResourceBindings?.requiredValidation
-        : 'form_filler.error_required';
-
-      const fieldNameReference = {
-        key: getFieldNameKey(textResourceBindings, undefined),
-        makeLowerCase: true,
-      };
-
-      validations.push({
-        message: {
-          key,
-          params: [fieldNameReference],
-        },
-        severity: 'error',
-        source: FrontendValidationSource.EmptyField,
-        category: ValidationMask.Required,
-      });
-    }
-    return validations;
+  useEmptyFieldValidation(node: LayoutNode<'List'>): ComponentValidation[] {
+    return useValidateListIsEmpty(node);
   }
 
   validateDataModelBindings(ctx: LayoutValidationCtx<'List'>): string[] {
