+++ src/layout/Subform/index.tsx
@@ -3,13 +3,7 @@ import { Route, Routes } from 'react-router-dom';
 import type { JSX, ReactNode } from 'react';
 
 import { TaskStoreProvider } from 'src/core/contexts/taskStoreContext';
-import {
-  type ComponentValidation,
-  FrontendValidationSource,
-  type SubformValidation,
-  type ValidationDataSources,
-  ValidationMask,
-} from 'src/features/validation';
+import { type ComponentValidation } from 'src/features/validation';
 import { type SummaryRendererProps } from 'src/layout/LayoutComponent';
 import { SubformDef } from 'src/layout/Subform/config.def.generated';
 import { SubformComponent } from 'src/layout/Subform/SubformComponent';
@@ -17,6 +11,7 @@ import { SubformValidator } from 'src/layout/Subform/SubformValidator';
 import { RedirectBackToMainForm, SubformForm, SubformWrapper } from 'src/layout/Subform/SubformWrapper';
 import { SubformSummaryComponent } from 'src/layout/Subform/Summary/SubformSummaryComponent';
 import { SubformSummaryComponent2 } from 'src/layout/Subform/Summary/SubformSummaryComponent2';
+import { useValidateSubform } from 'src/layout/Subform/useValidateSubform';
 import type { PropsFromGenericComponent, SubRouting, ValidateComponent } from 'src/layout';
 import type { NodeValidationProps } from 'src/layout/layout';
 import type { SubformSummaryOverrideProps } from 'src/layout/Summary2/config.generated';
@@ -73,69 +68,7 @@ export class Subform extends SubformDef implements ValidateComponent<'Subform'>,
     );
   }
 
-  runComponentValidation(
-    node: LayoutNode<'Subform'>,
-    {
-      applicationMetadata,
-      dataElementsSelector,
-      nodeDataSelector,
-      layoutSets,
-      dataElementHasErrorsSelector,
-    }: ValidationDataSources,
-  ): ComponentValidation[] {
-    const layoutSetName = nodeDataSelector((picker) => picker(node.id, 'Subform')?.layout.layoutSet, [node]);
-    if (!layoutSetName) {
-      throw new Error(`Layoutset not found for node with id ${node.id}.`);
-    }
-    const targetType = layoutSets.find((set) => set.id === layoutSetName)?.dataType;
-    if (!targetType) {
-      throw new Error(`Data type not found for layout with name ${layoutSetName}`);
-    }
-    const dataTypeDefinition = applicationMetadata.dataTypes.find((x) => x.id === targetType);
-    if (dataTypeDefinition === undefined) {
-      return [];
-    }
-
-    const validations: ComponentValidation[] = [];
-
-    const elements = dataElementsSelector((d) => d.filter((x) => x.dataType === targetType), [targetType]);
-    const numDataElements = Array.isArray(elements) ? elements.length : 0;
-    const { minCount, maxCount } = dataTypeDefinition;
-
-    if (minCount > 0 && numDataElements < minCount) {
-      validations.push({
-        message: { key: 'form_filler.error_min_count_not_reached_subform', params: [minCount, targetType] },
-        severity: 'error',
-        source: FrontendValidationSource.Component,
-        category: ValidationMask.Required,
-      });
-    }
-
-    if (maxCount > 0 && numDataElements > maxCount) {
-      validations.push({
-        message: { key: 'form_filler.error_max_count_reached_subform_local', params: [targetType, maxCount] },
-        severity: 'error',
-        source: FrontendValidationSource.Component,
-        category: ValidationMask.Required,
-      });
-    }
-
-    const subformIdsWithError = Array.isArray(elements)
-      ? elements?.map((dE) => dE.id).filter((id) => dataElementHasErrorsSelector(id))
-      : [];
-    if (subformIdsWithError?.length) {
-      const validation: SubformValidation = {
-        subformDataElementIds: subformIdsWithError,
-        message: { key: 'form_filler.error_validation_inside_subform', params: [targetType] },
-        severity: 'error',
-        source: FrontendValidationSource.Component,
-        category: ValidationMask.Required,
-        noIncrementalUpdates: true, // Validations for subform data is not updated incrementally in the main form
-      };
-
-      validations.push(validation);
-    }
-
-    return validations;
+  useComponentValidation(node: LayoutNode<'Subform'>): ComponentValidation[] {
+    return useValidateSubform(node);
   }
 }
