+++ src/layout/index.ts
@@ -3,7 +3,7 @@ import type { MutableRefObject, ReactNode } from 'react';
 import { getComponentConfigs } from 'src/layout/components.generated';
 import type { CompBehaviors } from 'src/codegen/Config';
 import type { DisplayData } from 'src/features/displayData';
-import type { BaseValidation, ComponentValidation, ValidationDataSources } from 'src/features/validation';
+import type { BaseValidation, ComponentValidation } from 'src/features/validation';
 import type { IDataModelReference } from 'src/layout/common.generated';
 import type { IGenericComponentProps } from 'src/layout/GenericComponent';
 import type { CompInternal, CompTypes } from 'src/layout/layout';
@@ -81,27 +81,27 @@ type TypeFromDef<Def extends CompDef> = Def extends CompDef<infer T> ? T : CompT
 export function implementsAnyValidation<Def extends CompDef>(
   def: Def,
 ): def is Def & (ValidateEmptyField<TypeFromDef<Def>> | ValidateComponent<TypeFromDef<Def>>) {
-  return 'runEmptyFieldValidation' in def || 'runComponentValidation' in def;
+  return 'useEmptyFieldValidation' in def || 'useComponentValidation' in def;
 }
 
 export interface ValidateEmptyField<Type extends CompTypes> {
-  runEmptyFieldValidation: (node: LayoutNode<Type>, validationContext: ValidationDataSources) => ComponentValidation[];
+  useEmptyFieldValidation: (node: LayoutNode<Type>) => ComponentValidation[];
 }
 
 export function implementsValidateEmptyField<Def extends CompDef>(
   def: Def,
 ): def is Def & ValidateEmptyField<TypeFromDef<Def>> {
-  return 'runEmptyFieldValidation' in def;
+  return 'useEmptyFieldValidation' in def;
 }
 
 export interface ValidateComponent<Type extends CompTypes> {
-  runComponentValidation: (node: LayoutNode<Type>, validationContext: ValidationDataSources) => ComponentValidation[];
+  useComponentValidation: (node: LayoutNode<Type>) => ComponentValidation[];
 }
 
 export function implementsValidateComponent<Def extends CompDef>(
   def: Def,
 ): def is Def & ValidateComponent<TypeFromDef<Def>> {
-  return 'runComponentValidation' in def;
+  return 'useComponentValidation' in def;
 }
 
 export interface SubRouting<Type extends CompTypes> {
